// ADW app with AEV/assets/source-code/adw-3.1.1-aev/libs/content-ee/process-services-extension/src/lib/models/process-service.model.ts
/*
 * Copyright © 2005 - 2021 Alfresco Software, Ltd. All rights reserved.
 *
 * License rights for this program may be obtained from Alfresco Software, Ltd.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

export interface UploadWidgetType {
    id: string;
    type: string;
    link: boolean;
}

export const ALL_APPS = 0;